turtlebot-agent/pyproject.toml
[project]
name = "turtlebot-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "grpcio>=1.74.0",
    "protobuf>=6.31.1",
]
